Updated content grammar
Edited content of "Adding custom icons to your site" of this page.
There is a grammar mistake in "To further enrich your site design, you can add references to custom icons in your metadata, and these will be displayed in certain contexts. The most commonly used of these is the favicon (short for "favorites icon", referring to its use in the "favorites" or "bookmarks" lists in browsers)."
I changed "commonly used" to "common use".